+++ services/users/project/api/routes/auth.py
@@ -2,7 +2,7 @@ from flask import Blueprint, jsonify
 from sqlalchemy import exc, or_
 
 from project import db, bcrypt
-from project.api.models import User
+from project.api.models.user import User
 from project.api.utils import authenticate
 from project.api.utils import post_request
