+++ services/users/project/tests/routes/test_auth.py
@@ -3,7 +3,8 @@ import json
 from flask import current_app
 
 from project import db
-from project.api.models import User
+from project.api.models.user import User
+from project.tests.base import BaseTestCase
 from project.tests.base import BaseTestCase
 from project.tests.utils import add_user
